add multiple input-fields via edge drag
