+++ Stitch/Graph/LayerInspector/GenericFlyoutView.swift
@@ -130,7 +130,7 @@ struct GenericFlyoutRowView: View {
     }
     
     @MainActor
-    var isPropertyRowSelected: Bool {
+    var isSelectedInspectorRow: Bool {
         graph.propertySidebar.selectedProperty == layerInspectorRowId
     }
     
@@ -160,7 +160,6 @@ struct GenericFlyoutRowView: View {
             if !layerInputObserver.port.isShadowInput {
                 // For the layer inspector row button, use a
                 LayerInspectorRowButton(graph: graph,
-                                        document: document,
                                         layerInputObserver: layerInputObserver,
                                         layerInspectorRowId: layerInspectorRowId,
                                         // For layer inspector row button, provide a NodeIOCoordinate that assumes unpacked + field index
@@ -168,6 +167,7 @@ struct GenericFlyoutRowView: View {
                                                                     nodeId: node.id),
                                          packedInputCanvasItemId: canvasItemId,
                                         isHovered: isHovered,
+                                        isSelectedInspectorRow: isSelectedInspectorRow,
                                         fieldIndex: fieldIndex)
             }
                                     
@@ -189,7 +189,7 @@ struct GenericFlyoutRowView: View {
                                 isFieldInMultifieldInput: isMultifield,
                                 isForFlyout: true,
                                 // Always false for flyout row
-                                isSelectedInspectorRow: isPropertyRowSelected,
+                                isSelectedInspectorRow: isSelectedInspectorRow,
                                 useIndividualFieldLabel: layerInputObserver.useIndividualFieldLabel(activeIndex:  document.activeIndex))
             }
             
@@ -237,126 +237,5 @@ struct FlyoutBackgroundColorModifier: ViewModifier {
                         }
                 }
             }
-    }
-}
-
-
-//extension GraphState {
-extension StitchDocumentViewModel {
-    @MainActor
-    func addLayerFieldToGraph(layerInput: LayerInputPort,
-                              nodeId: NodeId,
-                              fieldIndex: Int,
-                              groupNodeFocused: NodeId?) {
-        
-        let document = self
-        let graph = document.visibleGraph
-        
-        guard let node = graph.getNode(nodeId),
-              let layerNode = node.layerNode else {
-            log("LayerInputFieldAddedToGraph: no node, layer node and/or document")
-            fatalErrorIfDebug()
-            return
-        }
-        
-        let portObserver: LayerInputObserver = layerNode[keyPath: layerInput.layerNodeKeyPath]
-        
-        let previousPackMode = portObserver.mode
-        
-        guard let unpackedPort: InputLayerNodeRowData = portObserver._unpackedData.allPorts[safe: fieldIndex] else {
-            fatalErrorIfDebug("LayerInputFieldAddedToGraph: no unpacked port for fieldIndex \(fieldIndex)")
-            return
-        }
-        
-        
-        // MARK: CREATING AND INITIALIZING THE CANVAS ITEM VIEW MODEL ITSELF
-                
-        // First field-group grabbed since layers don't have differing groups within one input
-        guard let unpackedPortParentFieldGroupType: FieldGroupType = layerInput
-            .getDefaultValue(for: layerNode.layer)
-            .getNodeRowType(nodeIO: .input, layerInputPort: layerInput, isLayerInspector: true)
-            .fieldGroupTypes
-            .first else {
-            fatalErrorIfDebug()
-            return
-        }
-        
-        let activeIndex = document.activeIndex
-        
-        let canvasObserver = CanvasItemViewModel(
-            id: CanvasItemId.layerInput(LayerInputCoordinate(node: nodeId,
-                                                             keyPath: unpackedPort.id)),
-            position: document.newCanvasItemInsertionLocation,
-            zIndex: graph.highestZIndex + 1,
-            parentGroupNodeId: groupNodeFocused,
-            inputRowObservers: [unpackedPort.rowObserver],
-            outputRowObservers: [])
-        
-        canvasObserver.assignNodeReferenceAndUpdateFieldGroupsOnRowViewModels(
-            node,
-            activeIndex: activeIndex,
-            unpackedPortParentFieldGroupType: unpackedPortParentFieldGroupType,
-            unpackedPortIndex: fieldIndex,
-            graph: graph)
-        
-        unpackedPort.canvasObserver = canvasObserver
-        
-        
-
-        // MARK: Change the pack mode
-        
-        let newPackMode = portObserver.mode
-        if previousPackMode != newPackMode {
-            portObserver.wasPackModeToggled(document: document)
-        }
-        
-        
-        // MARK: RESET CACHE
-        
-        graph.resetLayerInputsCache(layerNode: layerNode,
-                                    activeIndex: activeIndex) // Why?
-    }
-}
-
-struct LayerInputFieldAddedToGraph: StitchDocumentEvent {
-    
-    let layerInput: LayerInputPort
-    let nodeId: NodeId
-    let fieldIndex: Int
-    
-    @MainActor
-    func handle(state: StitchDocumentViewModel) {
-        state.handleLayerInputFieldAddedToCanvas(
-            layerInput: layerInput,
-            nodeId: nodeId,
-            fieldIndex: fieldIndex)
-    }
-}
-
-extension StitchDocumentViewModel {
-    @MainActor
-    func handleLayerInputFieldAddedToCanvas(layerInput: LayerInputPort,
-                                            nodeId: NodeId,
-                                            fieldIndex: Int) {
-        let graph = self.visibleGraph
-        
-        let addLayerField = { (nodeId: NodeId) in
-            self.addLayerFieldToGraph(layerInput: layerInput,
-                                       nodeId: nodeId,
-                                       fieldIndex: fieldIndex,
-                                       groupNodeFocused: self.groupNodeFocused?.groupNodeId)
-        }
-        
-        if let multiselectInputs = graph.propertySidebar.inputsCommonToSelectedLayers,
-           let layerMultiselectInput = multiselectInputs.first(where: { $0 == layerInput}) {
-            
-            layerMultiselectInput.multiselectObservers(graph).forEach { observer in
-                addLayerField(observer.nodeId)
-            }
-        } else {
-            addLayerField(nodeId)
-        }
-        
-        self.encodeProjectInBackground()
     }
 }
