+++ Stitch/Graph/LayerInspector/LayerInspectorView.swift
@@ -334,24 +334,6 @@ struct LayerInspectorOutputsSectionView: View {
 }
 
 extension GraphState {
-    
-    // i.e. "layers focused in the inspector" are just the primarily-selected layers in the sidebar
-    @MainActor
-    var inspectorFocusedLayers: NodeIdSet {
-        self.layersSidebarViewModel.selectionState.primary
-    }
-    
-    @MainActor
-    var inspectorFocusedLayerNodes: LayerNodes {
-        self.layersSidebarViewModel.selectionState.primary
-            .reduce(into: LayerNodes()) { acc, id in
-                if let layerNode: LayerNodeViewModel = self.getLayerNode(id) {
-                    acc.append(layerNode)
-                }
-            }
-    }
-    
-    
     // Note: just used for `LayerInspectorView`
     @MainActor
     func getLayerInspectorData() -> (header: String,
